[TypeScript] Convert a few files
