+++ package.json
@@ -54,7 +54,7 @@
     "@artsy/express-reloadable": "1.4.0",
     "@artsy/palette": "2.21.1",
     "@artsy/passport": "1.1.0",
-    "@artsy/reaction": "8.0.2",
+    "@artsy/reaction": "8.0.3",
     "@artsy/stitch": "3.0.0",
     "@babel/core": "^7.0.0",
     "@babel/node": "^7.0.0",
@@ -238,6 +238,7 @@
     "@types/react-dom": "^16.0.7",
     "@types/styled-system": "^3.0.4",
     "@types/webpack": "^4.4.11",
+    "@types/webpack-env": "^1.13.6",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
     "benv": "^3.3.0",
