+++ src/desktop/apps/artist/client.tsx
@@ -1,20 +1,13 @@
 import { buildClientApp } from "reaction/Artsy/Router/client"
 import { data as sd } from "sharify"
 import { routes } from "reaction/Apps/Artist/routes"
-import mediator from "desktop/lib/mediator.coffee"
 import React from "react"
 import ReactDOM from "react-dom"
 import styled from "styled-components"
 import qs from "querystring"
 import { clone, isArray } from "underscore"
 
-mediator.on("artist:filter:changed", filters => {
-  onFilterChange(filters)
-})
-
-mediator.on("artist:tabclick", ({ to }) => {
-  window.analytics.page({ path: to }, { integrations: { Marketo: false } })
-})
+const mediator = require("desktop/lib/mediator.coffee")
 
 buildClientApp({
   routes,
@@ -39,6 +32,14 @@ buildClientApp({
 if (module.hot) {
   module.hot.accept()
 }
+
+mediator.on("artist:filter:changed", filters => {
+  onFilterChange(filters)
+})
+
+mediator.on("artist:tabclick", ({ to }) => {
+  window.analytics.page({ path: to }, { integrations: { Marketo: false } })
+})
 
 // Update URL with current filters and sort.
 const onFilterChange = filters => {
