+++ src/desktop/index.js
@@ -16,12 +16,12 @@ app.use(require("./apps/apply"))
 app.use(require("./apps/auctions"))
 app.use(require("./apps/auctions2"))
 app.use(require("./apps/auction_lots"))
-app.use(require("./apps/artist/server"))
+app.use(require("./apps/artist/server").app)
 app.use(require("./apps/artists"))
 app.use(require("./apps/auction"))
 app.use(require("./apps/auction_support"))
 app.use(require("./apps/artwork"))
-app.use(require("./apps/artwork2/server"))
+app.use(require("./apps/artwork2/server").app)
 app.use(require("./apps/about"))
 app.use(require("./apps/collect_art"))
 app.use(require("./apps/collect2/server"))
