+++ src/typings/express.d.ts
@@ -1,0 +1,7 @@
+import * as express from "express"
+
+declare module "express" {
+  interface Request {
+    user?: any
+  }
+}
