+++ src/typings/global.d.ts
@@ -1,0 +1,5 @@
+interface Window {
+  analytics?: {
+    page: (object, object) => void
+  }
+}
